fixbug: 修复面额张数截断与解析边界

+++ un260/counting/counting_data_types.h
@@ -8,7 +8,7 @@
 
 typedef struct {
     int value;
-    uint8_t pcs;
+    uint16_t pcs;
     float amount;
 } denom_t;
